Reorganize file locations.
