+++ Menus/MenuBuilder.cs
@@ -1,10 +1,11 @@
-﻿using System;
+﻿using ADOnetSakilaKoppling.Services;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ADOnetSakilaKoppling.Services
+namespace ADOnetSakilaKoppling.Menus
 {
     internal static class MenuBuilder
     {
