+++ Menus/MenuOption.cs
@@ -5,7 +5,7 @@ using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ADOnetSakilaKoppling.Services
+namespace ADOnetSakilaKoppling.Menus
 {
     internal class MenuOption
     {
